Renamed DropdownClick component
